+++ backend/dcl/src/job_end/mod.rs
@@ -14,7 +14,7 @@ use tokio::sync::RwLock;
 
 use crate::node_end::NodePool;
 use crate::DatasetPair;
-use messages::ClientMessage;
+use messages::{ClientMessage, ReadLengthPrefix, WriteLengthPrefix};
 use models::predictions::Prediction;
 
 use utils::anon::{anonymise_dataset, deanonymise_dataset};
