+++ backend/messages/Cargo.toml
@@ -14,3 +14,4 @@ serde_json = "1.0"
 log = "0.4"
 utils = { path = "../utils" }
 mongodb = { git = "https://github.com/alexander-jackson/mongo-rust-driver" }
+async-trait = "0.1.42"
